finish&fix
