+++ jest.config.js
@@ -1,4 +1,9 @@
 module.exports = {
+  "globals": {
+    'ts-jest': {
+      "tsConfigFile": 'tsconfig.json'
+    }
+  },
   "transform": {
     ".(ts|tsx)": "ts-jest"
   },
@@ -23,5 +28,6 @@ module.exports = {
   },
   "collectCoverageFrom": [
     "src/*.{js,ts}"
-  ]
+  ],
+  "testEnvironment": "jsdom"
 }
