+++ lib/createBrowserHistory.d.ts
@@ -1,0 +1,12 @@
+/**
+ * Creates and returns a history object that uses HTML5's history API
+ * (pushState, replaceState, and the popstate event) to manage history.
+ * This is the recommended method of managing history in browsers because
+ * it provides the cleanest URLs.
+ *
+ * Note: In browsers that do not support the HTML5 history API full
+ * page reloads will be used to preserve clean URLs. You can force this
+ * behavior using { forceRefresh: true } in options.
+ */
+declare const createBrowserHistory: CH.Browser.CreateHistory;
+export default createBrowserHistory;
