+++ lib/createHashHistory.d.ts
@@ -1,0 +1,2 @@
+declare const createHashHistory: CH.Hash.CreateHistory;
+export default createHashHistory;
