+++ lib/createHistory.d.ts
@@ -1,0 +1,2 @@
+declare const createHistory: CH.History.CreateHistory;
+export default createHistory;
