+++ lib/createMemoryHistory.d.ts
@@ -1,0 +1,2 @@
+declare const createMemoryHistory: CH.Memory.CreateHistory;
+export default createMemoryHistory;
