+++ lib/index.d.ts
@@ -1,0 +1,13 @@
+export declare const locationsAreEqual: import("./utils/type").Location.LocationsAreEqual;
+import Actions from './utils/Actions';
+declare const _default: {
+    locationsAreEqual: import("./utils/type").Location.LocationsAreEqual;
+    createHistory: import(".").Browser.CreateHistory;
+    createHashHistory: import(".").Hash.CreateHistory;
+    createMemoryHistory: import(".").Memory.CreateHistory;
+    useBasename: import(".").Basename.UseBasename;
+    useBeforeUnload: import(".").BeforeUnload.UseBeforeUnload;
+    useQueries: import(".").Queries.useQueries;
+    Actions: typeof Actions;
+};
+export default _default;
