+++ lib/useBasename.d.ts
@@ -1,0 +1,2 @@
+declare const useBasename: CH.Basename.UseBasename;
+export default useBasename;
