+++ lib/useBeforeUnload.d.ts
@@ -1,0 +1,7 @@
+/**
+ * Returns a new createHistory function that can be used to create
+ * history objects that know how to use the beforeunload event in web
+ * browsers to cancel navigation.
+ */
+declare const useBeforeUnload: CH.BeforeUnload.UseBeforeUnload;
+export default useBeforeUnload;
