+++ lib/useQueries.d.ts
@@ -1,0 +1,6 @@
+/**
+ * Returns a new createHistory function that may be used to create
+ * history objects that know how to handle URL queries.
+ */
+declare const useQueries: CH.Queries.useQueries;
+export default useQueries;
