+++ lib/utils/BrowserProtocol.d.ts
@@ -1,0 +1,6 @@
+export declare const getCurrentLocation: Utils.Browser.GetBrowserCurrentLocation;
+export declare const getUserConfirmation: Utils.Browser.GetUserConfirmation;
+export declare const startListener: Utils.Browser.StartListener;
+export declare const pushLocation: Utils.Browser.PushLocation;
+export declare const replaceLocation: Utils.Browser.ReplaceLocation;
+export declare const go: Utils.Browser.Go;
