+++ lib/utils/DOMStateStorage.d.ts
@@ -1,0 +1,3 @@
+export declare const createKey: Utils.DOMStateStorage.CreateKey;
+export declare const saveState: Utils.DOMStateStorage.SaveKey;
+export declare const readState: Utils.DOMStateStorage.ReadState;
