+++ lib/utils/ExecutionEnvironment.d.ts
@@ -1,0 +1,1 @@
+export declare const canUseDOM: boolean;
