+++ lib/utils/HashProtocol.d.ts
@@ -1,0 +1,6 @@
+export declare let getUserConfirmation: import("./type").Browser.GetUserConfirmation;
+export declare let go: import("./type").Browser.Go;
+export declare const getCurrentLocation: Utils.Hash.GetCurrentLocation;
+export declare const startListener: Utils.Hash.StartListener;
+export declare const pushLocation: Utils.Hash.PushLocation;
+export declare const replaceLocation: Utils.Hash.ReplaceLocation;
