+++ lib/utils/LocationUtils.d.ts
@@ -1,0 +1,4 @@
+export declare const createQuery: Utils.Location.CreateQuery;
+export declare const createLocation: Utils.Location.CreateLocation;
+export declare const statesAreEqual: Utils.Location.StatesAreEqual;
+export declare const locationsAreEqual: Utils.Location.LocationsAreEqual;
