+++ lib/utils/PathUtils.d.ts
@@ -1,0 +1,5 @@
+export declare const addQueryStringValueToPath: Utils.Path.AddQueryStringValueToPath;
+export declare const stripQueryStringValueFromPath: Utils.Path.StripQueryStringValueFromPath;
+export declare const getQueryStringValueFromPath: Utils.Path.GetQueryStringValueFromPath;
+export declare const parsePath: Utils.Path.ParsePath;
+export declare const createPath: Utils.Path.CreatePath;
