+++ lib/utils/RefreshProtocol.d.ts
@@ -1,0 +1,5 @@
+export declare let getUserConfirmation: import("./type").Browser.GetUserConfirmation;
+export declare let go: import("./type").Browser.Go;
+export declare const getCurrentLocation: Utils.Refresh.GetCurrentLocation;
+export declare const pushLocation: Utils.Refresh.PushLocation;
+export declare const replaceLocation: Utils.Refresh.ReplaceLocation;
