+++ lib/utils/runTransitionHook.d.ts
@@ -1,0 +1,2 @@
+declare const runTransitionHook: Utils.Transition.RunTransitionHook;
+export default runTransitionHook;
