+++ src/BrowserProtocol.ts
@@ -1,9 +1,8 @@
-import Utils from './type'
-
 import { createLocation } from './LocationUtils'
 import { addEventListener, removeEventListener } from './DOMUtils'
 import { saveState, readState } from './DOMStateStorage'
 import { createPath } from './PathUtils'
+import { Utils } from './index'
 
 const PopStateEvent = 'popstate'
