+++ src/HashProtocol.ts
@@ -13,7 +13,7 @@ import {
   getUserConfirmation as _getUserConfirmation,
   go as _go
 } from './BrowserProtocol'
-import Utils from './type'
+import { Utils } from './index'
 
 export let getUserConfirmation = _getUserConfirmation
 export let go = _go
