+++ tsconfig.json
@@ -3,6 +3,7 @@
     "moduleResolution": "node",
     "target": "es5",
     "module":"es2015",
+    "diagnostics": true,
     "lib": ["es2015", "es2016", "es2017", "dom", "esnext"],
     "jsx": "react",
     "strict": false,
@@ -13,6 +14,7 @@
     "allowSyntheticDefaultImports": true,
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
+    // "emitDeclarationOnly": true,
     "declarationDir": "dist/types",
     "outDir": "dist/lib",
     "typeRoots": [
